+++ angular/CHANGELOG.md
@@ -66,9 +66,14 @@
 *   [#1500][]: Configuring a provider with `FactoryProvider(Foo, null)` is now
     a compile-time error, instead of a misleading runtime error.
 
-*   [#1592][]: Using `@GeneratedInjector` with a `ValueProvider` bound to a
+*   [#1591][]: Using `@GenerateInjector` with a `ValueProvider` bound to a
     `String` instance that expects a _raw_ string (i.e `r'$5.00'`) no longer
     generates invalid code. Now _all_ strings are emitted as _raw_.
+
+*   [#1598][]: Using `@GenerateInjector` with a `ValueProvider` whose value
+    is created as a `const` object with _named_ arguments is now created
+    correctly. Before, all named arguments were skipped (left to default values,
+    which was often `null`).
 
 [#434]: https://github.com/dart-lang/angular/issues/434
 [#880]: https://github.com/dart-lang/angular/issues/880
@@ -79,7 +84,8 @@
 [#1540]: https://github.com/dart-lang/angular/issues/1540
 [#1558]: https://github.com/dart-lang/angular/issues/1558
 [#1570]: https://github.com/dart-lang/angular/issues/1570
-[#1592]: https://github.com/dart-lang/angular/pull/1592
+[#1591]: https://github.com/dart-lang/angular/issues/1591
+[#1598]: https://github.com/dart-lang/angular/issues/1598
 
 ### Other improvements